per:加入hungarian_tracker代价矩阵打分机制，弃用yolo内部tracker,由hungarian维护全部tracker，将iou与欧式距离与身份标签投票权重调整，对画面检测框绘制做出调整使其更加清楚，下一步预计参考HKUST，去掉bbox_kalman的预测运动阻尼，按照匀速直线预测。

+++ src/hnurm_radar/hnurm_radar/shared/type.py
@@ -88,4 +88,5 @@ class RobotState:
     last_seen_time: float = 0.0 # 上次真实观测的绝对时间戳
     
     # 兵种身份投票池: {label_name: count}
-    vote_pool: Dict[str, int] = field(default_factory=dict)
+    # // tunning: 将 count (int) 升级为 score (float)，以支持基于检测置信度的加权投票。
+    vote_pool: Dict[str, float] = field(default_factory=dict)
